Address Céline's comments
