+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/fragments/HomeFragment.java
@@ -59,8 +59,6 @@ public final class HomeFragment extends Fragment {
 
     setupListAdapter();
     setupListUpdates();
-
-    //    mHomeViewModel.setupDummyLAO();
   }
 
   private void setupListUpdates() {
