+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/model/Lao.java
@@ -1,6 +1,5 @@
 package com.github.dedis.student20_pop.model;
 
-import com.github.dedis.student20_pop.model.event.Event;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
@@ -126,131 +125,5 @@ public final class Lao {
 
   public void setRollCalls(Map<String, RollCall> rollCalls) {
     this.rollCalls = rollCalls;
-  }
-
-  public static class PendingUpdate implements Comparable<PendingUpdate> {
-
-    private long modificationTime;
-    private String messageId;
-
-    public PendingUpdate(long modificationTime, String messageId) {
-      this.modificationTime = modificationTime;
-      this.messageId = messageId;
-    }
-
-    public long getModificationTime() {
-      return modificationTime;
-    }
-
-    @Override
-    public int compareTo(PendingUpdate o) {
-      return Long.compare(modificationTime, o.modificationTime);
-    }
-  }
-
-  public static class RollCall extends Event {
-
-    private String id;
-    private String name;
-    private long creation;
-    private long start;
-    private long scheduled;
-    private long end;
-    private Set<String> attendees;
-
-    private String location;
-    private String description;
-
-    public RollCall() {
-      this.attendees = new HashSet<>();
-    }
-
-    public String getId() {
-      return id;
-    }
-
-    public void setId(String id) {
-      this.id = id;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public long getCreation() {
-      return creation;
-    }
-
-    public void setCreation(long creation) {
-      this.creation = creation;
-    }
-
-    public long getStart() {
-      return start;
-    }
-
-    public void setStart(long start) {
-      this.start = start;
-    }
-
-    public long getScheduled() {
-      return scheduled;
-    }
-
-    public void setScheduled(long scheduled) {
-      this.scheduled = scheduled;
-    }
-
-    public long getEnd() {
-      return end;
-    }
-
-    public void setEnd(long end) {
-      this.end = end;
-    }
-
-    public Set<String> getAttendees() {
-      return attendees;
-    }
-
-    public void setAttendees(Set<String> attendees) {
-      this.attendees = attendees;
-    }
-
-    public String getLocation() {
-      return location;
-    }
-
-    public void setLocation(String location) {
-      this.location = location;
-    }
-
-    public String getDescription() {
-      return description;
-    }
-
-    public void setDescription(String description) {
-      this.description = description;
-    }
-
-    @Override
-    public long getStartTimestamp() {
-      if (start != 0) {
-        return start;
-      }
-      return scheduled;
-    }
-
-    @Override
-    public long getEndTimestamp() {
-      if (end == 0) {
-        return Long.MAX_VALUE;
-      }
-      return end;
-    }
   }
 }
