+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/model/data/LAORepository.java
@@ -4,8 +4,8 @@ import android.util.Base64;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import com.github.dedis.student20_pop.model.Lao;
-import com.github.dedis.student20_pop.model.Lao.PendingUpdate;
-import com.github.dedis.student20_pop.model.Lao.RollCall;
+import com.github.dedis.student20_pop.model.PendingUpdate;
+import com.github.dedis.student20_pop.model.RollCall;
 import com.github.dedis.student20_pop.model.network.GenericMessage;
 import com.github.dedis.student20_pop.model.network.answer.Answer;
 import com.github.dedis.student20_pop.model.network.answer.Error;
@@ -311,7 +311,7 @@ public class LAORepository {
     Lao lao = laoById.get(channel).getLao();
     Log.d(TAG, "handleCreateRollCall: " + channel + " name " + createRollCall.getName());
 
-    Lao.RollCall rollCall = new Lao.RollCall();
+    RollCall rollCall = new RollCall();
     rollCall.setId(createRollCall.getId());
     rollCall.setCreation(createRollCall.getCreation());
